requirements.txt
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.8.5
sqlparse==0.3.1 
PyJWT==1.7.1
djangorestframework-simplejwt==4.3.0
Django==3.1.1
django-filter==2.3.0
djangorestframework==3.11.1
importlib-metadata==1.7.0
Markdown==3.2.2
Pillow==7.2.0
pytz==2020.1
sorl-thumbnail==12.6.3
sqlparse==0.3.1
zipp==3.1.0
